# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Lumen",
        "version": "1.0.0"
    },
    "paths": {
        "/ckan-datasets": {
            "get": {
                "summary": "Obtiene conjuntos de datos",
                "description": "Recupera una lista de conjuntos de datos desde CKAN, con la opción de paginación.",
                "operationId": "acb3f8f27446fc3893f2c98535541d25",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "description": "Índice de inicio para la paginación de los conjuntos de datos",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "rows",
                        "in": "query",
                        "description": "Número de filas o conjuntos de datos a devolver",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "count": {
                                            "description": "Número total de conjuntos de datos disponibles",
                                            "type": "integer"
                                        },
                                        "datasets": {
                                            "description": "Lista de conjuntos de datos",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Dataset"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud"
                    }
                }
            }
        },
        "/ckan-datasets/details/{id}": {
            "get": {
                "summary": "Obtiene detalles de un conjunto de datos específico",
                "description": "Recupera los detalles completos de un conjunto de datos específico en CKAN, incluyendo recursos y etiquetas.",
                "operationId": "3980342247beb8a2c95e4eb89d81e681",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID único del conjunto de datos",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del conjunto de datos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DatasetDetail"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud"
                    },
                    "404": {
                        "description": "Conjunto de datos no encontrado"
                    }
                }
            }
        },
        "/ckan-datasets-by-tag/{tag}": {
            "get": {
                "summary": "Obtiene conjuntos de datos por etiqueta",
                "description": "Recupera conjuntos de datos asociados con una etiqueta específica, con opciones de paginación.",
                "operationId": "7db2e396c0a76e7475d58a2db595fb99",
                "parameters": [
                    {
                        "name": "tag",
                        "in": "path",
                        "description": "Etiqueta para filtrar los conjuntos de datos",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start",
                        "in": "query",
                        "description": "Índice de inicio para la paginación de los conjuntos de datos",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "rows",
                        "in": "query",
                        "description": "Número de conjuntos de datos a devolver en la respuesta",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de conjuntos de datos filtrados por etiqueta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "count": {
                                            "description": "Número total de conjuntos de datos encontrados",
                                            "type": "integer"
                                        },
                                        "datasets": {
                                            "description": "Conjuntos de datos que coinciden con la etiqueta especificada",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Dataset"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud"
                    }
                }
            }
        },
        "/ckan-datasets/most_downloaded/{rows}": {
            "get": {
                "summary": "Obtiene los conjuntos de datos más descargados",
                "description": "Recupera una lista de los conjuntos de datos más descargados, limitada a un número específico de filas.",
                "operationId": "89b17f39548cad494eb5b09576596034",
                "parameters": [
                    {
                        "name": "rows",
                        "in": "path",
                        "description": "Número de filas o conjuntos de datos a devolver",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de los conjuntos de datos más descargados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "count": {
                                            "description": "Número total de conjuntos de datos más descargados disponibles",
                                            "type": "integer"
                                        },
                                        "datasets": {
                                            "description": "Conjuntos de datos más descargados",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Dataset"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud"
                    }
                }
            }
        },
        "/ckan-datasets/latest/{rows}": {
            "get": {
                "summary": "Obtiene los conjuntos de datos más recientes",
                "description": "Recupera una lista de los conjuntos de datos más recientemente actualizados, limitada a un número específico de filas.",
                "operationId": "251999623dab04b0938af8274f9a1790",
                "parameters": [
                    {
                        "name": "rows",
                        "in": "path",
                        "description": "Número de filas o conjuntos de datos a devolver",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de los conjuntos de datos más recientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "count": {
                                            "description": "Número total de conjuntos de datos más recientes disponibles",
                                            "type": "integer"
                                        },
                                        "datasets": {
                                            "description": "Conjuntos de datos más recientes",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Dataset"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Dataset": {
                "title": "Conjunto de Datos",
                "description": "Representa un conjunto de datos en CKAN",
                "properties": {
                    "id": {
                        "description": "ID único del conjunto de datos",
                        "type": "string"
                    },
                    "name": {
                        "description": "Nombre del conjunto de datos",
                        "type": "string"
                    },
                    "notes": {
                        "description": "Descripción o notas sobre el conjunto de datos",
                        "type": "string"
                    },
                    "resource_formats": {
                        "description": "Formatos de los recursos disponibles para el conjunto de datos",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "api_detail_url": {
                        "description": "URL de la API para obtener detalles del conjunto de datos",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}